Separate GenericConstructor.cs
