+++ Assets/VRM/UniJSON/Scripts/GenericConstructor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6121925cab5acc54bbb0e0e213783394
+timeCreated: 1549447177
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
